@@ -1,42 +1,23 @@
 name: example
-description: A new Flutter project.
-
+description: A project that showcases usage of cached_network_image
+publish_to: none
+version: 1.0.0+1
 environment:
-  sdk: ">=2.18.0 <3.0.0"
-  flutter: ">=3.3.0"
+  sdk: ^3.0.0
+  flutter: '>=3.10.0'
 
 dependencies:
-  flutter:
-    sdk: flutter
+  baseflow_plugin_template: ^2.1.2
   cached_network_image:
     path: ../
-  baseflow_plugin_template:
-    git:
-      url: https://github.com/Baseflow/baseflow_plugin_template.git
-      ref: 922aa52773308bec29f0cfa2661b43ecdedd50c5
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  flutter:
+    sdk: flutter
+  flutter_blurhash: ^0.8.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.4
+  flutter_lints: ^2.0.3
 
-
-# For information on the generic Dart part of this file, see the
-# following page: https://www.dartlang.org/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
-
-  # To add assets to your application, add an assets section, like this:
-  assets:
-    - packages/baseflow_plugin_template/logo.png
-    - packages/baseflow_plugin_template/poweredByBaseflow.png
