@@ -2,27 +2,30 @@ name: cached_network_image
 description: Flutter library to load and cache network images.
   Can also be used with placeholder and error widgets.
 homepage: https://github.com/Baseflow/flutter_cached_network_image
-version: 3.2.3
+topics:
+  - cache
+  - image
+  - network-image
+version: 3.3.1
+environment:
+  sdk: ^3.0.0
+  flutter: '>=3.10.0'
 
 dependencies:
-  flutter:
-    sdk: flutter
-  flutter_cache_manager: ^3.3.0
-  octo_image: ^1.0.0
+  cached_network_image_platform_interface: ^4.0.0
   cached_network_image_web:
     git:
       url: git@github.com:Insight-Timer/flutter_cached_network_image.git
       ref: feature/errorListener
       path: cached_network_image_web
-
+  flutter:
+    sdk: flutter
+  flutter_cache_manager: ^3.3.1
+  octo_image: ^2.0.0
 
 dev_dependencies:
+  file: '>=6.1.4 <8.0.0'
+  flutter_lints: ^2.0.3
   flutter_test:
     sdk: flutter
-  mocktail: ^0.1.1
-  flutter_lints: ^1.0.4
-  file: ^6.1.0
-
-environment:
-  flutter: ">=3.3.0"
-  sdk: ">=2.18.0 <3.0.0"
+  mocktail: ^1.0.0
