@@ -34,7 +34,7 @@ typedef ProgressIndicatorBuilder = Widget Function(
 typedef LoadingErrorWidgetBuilder = Widget Function(
   BuildContext context,
   String url,
-  dynamic error,
+  Object error,
 );
 
 /// Image widget to show NetworkImage with caching functionality.
@@ -50,20 +50,20 @@ class CachedNetworkImage extends StatelessWidget {
   /// [BaseCacheManager] as the in memory [ImageCache] of the [ImageProvider].
   /// [url] is used by both the disk and memory cache. The scale is only used
   /// to clear the image from the [ImageCache].
-  static Future evictFromCache(
+  static Future<bool> evictFromCache(
     String url, {
     String? cacheKey,
     BaseCacheManager? cacheManager,
-    double scale = 1.0,
+    double scale = 1,
   }) async {
-    cacheManager = cacheManager ?? DefaultCacheManager();
-    await cacheManager.removeFile(cacheKey ?? url);
+    final effectiveCacheManager = cacheManager ?? DefaultCacheManager();
+    await effectiveCacheManager.removeFile(cacheKey ?? url);
     return CachedNetworkImageProvider(url, scale: scale).evict();
   }
 
   final CachedNetworkImageProvider _image;
 
-  /// Option to use cachemanager with other settings
+  /// Option to use cacheManager with other settings
   final BaseCacheManager? cacheManager;
 
   /// The target image that is displayed.
@@ -211,7 +211,7 @@ class CachedNetworkImage extends StatelessWidget {
   /// loaded image. Next to that it supports most features of a default Image
   /// widget.
   CachedNetworkImage({
-    Key? key,
+    super.key,
     required this.imageUrl,
     this.httpHeaders,
     this.imageBuilder,
@@ -242,7 +242,7 @@ class CachedNetworkImage extends StatelessWidget {
     this.errorListener,
     ImageRenderMethodForWeb imageRenderMethodForWeb =
         ImageRenderMethodForWeb.HtmlImage,
-  })  : _image = CachedNetworkImageProvider(
+  }) : _image = CachedNetworkImageProvider(
           imageUrl,
           headers: httpHeaders,
           cacheManager: cacheManager,
@@ -258,10 +258,10 @@ class CachedNetworkImage extends StatelessWidget {
   Widget build(BuildContext context) {
     var octoPlaceholderBuilder =
         placeholder != null ? _octoPlaceholderBuilder : null;
-    var octoProgressIndicatorBuilder =
+    final octoProgressIndicatorBuilder =
         progressIndicatorBuilder != null ? _octoProgressIndicatorBuilder : null;
 
-    ///If there is no placeholer OctoImage does not fade, so always set an
+    ///If there is no placeholder OctoImage does not fade, so always set an
     ///(empty) placeholder as this always used to be the behaviour of
     ///CachedNetworkImage.
     if (octoPlaceholderBuilder == null &&
@@ -314,7 +314,10 @@ class CachedNetworkImage extends StatelessWidget {
       downloaded = progress.cumulativeBytesLoaded;
     }
     return progressIndicatorBuilder!(
-        context, imageUrl, DownloadProgress(imageUrl, totalSize, downloaded));
+      context,
+      imageUrl,
+      DownloadProgress(imageUrl, totalSize, downloaded),
+    );
   }
 
   Widget _octoErrorBuilder(
