@@ -7,7 +7,7 @@ import 'package:flutter/scheduler.dart';
 
 /// Slows down animations by this factor to help in development.
 double get timeDilation => _timeDilation;
-double _timeDilation = 1.0;
+double _timeDilation = 1;
 
 /// An ImageStreamCompleter with support for loading multiple images.
 class MultiImageStreamCompleter extends ImageStreamCompleter {
@@ -22,25 +22,28 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
     InformationCollector? informationCollector,
   })  : _informationCollector = informationCollector,
         _scale = scale {
-    codec.listen((event) {
-      if (_timer != null) {
-        _nextImageCodec = event;
-      } else {
-        _handleCodecReady(event);
-      }
-    }, onError: (dynamic error, StackTrace stack) {
-      reportError(
-        context: ErrorDescription('resolving an image codec'),
-        exception: error,
-        stack: stack,
-        informationCollector: informationCollector,
-        silent: true,
-      );
-    });
+    codec.listen(
+      (event) {
+        if (_timer != null) {
+          _nextImageCodec = event;
+        } else {
+          _handleCodecReady(event);
+        }
+      },
+      onError: (Object error, StackTrace stack) {
+        reportError(
+          context: ErrorDescription('resolving an image codec'),
+          exception: error,
+          stack: stack,
+          informationCollector: informationCollector,
+          silent: true,
+        );
+      },
+    );
     if (chunkEvents != null) {
       _chunkSubscription = chunkEvents.listen(
         reportImageChunkEvent,
-        onError: (dynamic error, StackTrace stack) {
+        onError: (Object error, StackTrace stack) {
           reportError(
             context: ErrorDescription('loading an image'),
             exception: error,
@@ -58,10 +61,13 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
   final double _scale;
   final InformationCollector? _informationCollector;
   ui.FrameInfo? _nextFrame;
+
   // When the current was first shown.
   Duration? _shownTimestamp;
+
   // The requested duration for the current frame;
   Duration? _frameDuration;
+
   // How many frames have been emitted so far.
   int _framesEmitted = 0;
   Timer? _timer;
@@ -125,7 +131,7 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
   Future<void> _decodeNextFrameAndSchedule() async {
     try {
       _nextFrame = await _codec!.getNextFrame();
-    } catch (exception, stack) {
+    } on Object catch (exception, stack) {
       reportError(
         context: ErrorDescription('resolving an image frame'),
         exception: exception,
