@@ -2,23 +2,29 @@
 // Use of this source code is governed by a MIT-style license that can be
 // found in the LICENSE file.
 
+import 'dart:async';
 import 'dart:typed_data';
 
 import 'package:file/memory.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:mocktail/mocktail.dart';
-import 'dart:async';
 
 class FakeCacheManager extends Mock implements CacheManager {
   void throwsNotFound(String url) {
-    when(() => getFileStream(
-              url,
-              key: any(named: 'key'),
-              headers: any(named: 'headers'),
-              withProgress: any(named: 'withProgress'),
-            ))
-        .thenThrow(HttpExceptionWithStatus(404, 'Invalid statusCode: 404',
-            uri: Uri.parse(url)));
+    when(
+      () => getFileStream(
+        url,
+        key: any(named: 'key'),
+        headers: any(named: 'headers'),
+        withProgress: any(named: 'withProgress'),
+      ),
+    ).thenThrow(
+      HttpExceptionWithStatus(
+        404,
+        'Invalid statusCode: 404',
+        uri: Uri.parse(url),
+      ),
+    );
   }
 
   ExpectedData returns(
@@ -32,17 +38,21 @@ class FakeCacheManager extends Mock implements CacheManager {
         Uint8List.fromList(imageData.skip(offset).take(chunkSize).toList()),
     ];
 
-    when(() => getFileStream(
-          url,
-          key: any(named: 'key'),
-          headers: any(named: 'headers'),
-          withProgress: any(named: 'withProgress'),
-        )).thenAnswer((_) => _createResultStream(
-          url,
-          chunks,
-          imageData,
-          delayBetweenChunks,
-        ));
+    when(
+      () => getFileStream(
+        url,
+        key: any(named: 'key'),
+        headers: any(named: 'headers'),
+        withProgress: any(named: 'withProgress'),
+      ),
+    ).thenAnswer(
+      (_) => _createResultStream(
+        url,
+        chunks,
+        imageData,
+        delayBetweenChunks,
+      ),
+    );
 
     return ExpectedData(
       chunks: chunks.length,
@@ -57,19 +67,23 @@ class FakeCacheManager extends Mock implements CacheManager {
     List<int> imageData,
     Duration? delayBetweenChunks,
   ) async* {
-    var totalSize = imageData.length;
+    final totalSize = imageData.length;
     var downloaded = 0;
-    for (var chunk in chunks) {
+    for (final chunk in chunks) {
       downloaded += chunk.length;
       if (delayBetweenChunks != null) {
-        await Future.delayed(delayBetweenChunks);
+        await Future<void>.delayed(delayBetweenChunks);
       }
       yield DownloadProgress(url, totalSize, downloaded);
     }
-    var file = MemoryFileSystem().systemTempDirectory.childFile('test.jpg');
+    final file = MemoryFileSystem().systemTempDirectory.childFile('test.jpg');
     await file.writeAsBytes(imageData);
-    yield FileInfo(file, FileSource.Online,
-        DateTime.now().add(const Duration(days: 1)), url);
+    yield FileInfo(
+      file,
+      FileSource.Online,
+      DateTime.now().add(const Duration(days: 1)),
+      url,
+    );
   }
 }
 
@@ -85,19 +99,23 @@ class FakeImageCacheManager extends Mock implements ImageCacheManager {
         Uint8List.fromList(imageData.skip(offset).take(chunkSize).toList()),
     ];
 
-    when(() => getImageFile(
-          url,
-          key: any(named: 'key'),
-          headers: any(named: 'headers'),
-          withProgress: any(named: 'withProgress'),
-          maxHeight: any(named: 'maxHeight'),
-          maxWidth: any(named: 'maxWidth'),
-        )).thenAnswer((_) => _createResultStream(
-          url,
-          chunks,
-          imageData,
-          delayBetweenChunks,
-        ));
+    when(
+      () => getImageFile(
+        url,
+        key: any(named: 'key'),
+        headers: any(named: 'headers'),
+        withProgress: any(named: 'withProgress'),
+        maxHeight: any(named: 'maxHeight'),
+        maxWidth: any(named: 'maxWidth'),
+      ),
+    ).thenAnswer(
+      (_) => _createResultStream(
+        url,
+        chunks,
+        imageData,
+        delayBetweenChunks,
+      ),
+    );
 
     return ExpectedData(
       chunks: chunks.length,
@@ -112,19 +130,23 @@ class FakeImageCacheManager extends Mock implements ImageCacheManager {
     List<int> imageData,
     Duration? delayBetweenChunks,
   ) async* {
-    var totalSize = imageData.length;
+    final totalSize = imageData.length;
     var downloaded = 0;
-    for (var chunk in chunks) {
+    for (final chunk in chunks) {
       downloaded += chunk.length;
       if (delayBetweenChunks != null) {
-        await Future.delayed(delayBetweenChunks);
+        await Future<void>.delayed(delayBetweenChunks);
       }
       yield DownloadProgress(url, totalSize, downloaded);
     }
-    var file = MemoryFileSystem().systemTempDirectory.childFile('test.jpg');
+    final file = MemoryFileSystem().systemTempDirectory.childFile('test.jpg');
     await file.writeAsBytes(imageData);
-    yield FileInfo(file, FileSource.Online,
-        DateTime.now().add(const Duration(days: 1)), url);
+    yield FileInfo(
+      file,
+      FileSource.Online,
+      DateTime.now().add(const Duration(days: 1)),
+      url,
+    );
   }
 }
 
