@@ -19,110 +19,150 @@ void main() {
     PaintingBinding.instance.imageCache.clearLiveImages();
   });
 
-  test('Supplying an ImageCacheManager should call getImageFile', () async {
-    var url = 'foo.nl';
+  test(
+    'Supplying an ImageCacheManager should call getImageFile',
+    () async {
+      var url = 'foo.nl';
 
-    var cacheManager = FakeImageCacheManager();
-    cacheManager.returns(url, kTransparentImage);
-    final imageAvailable = Completer<void>();
+      var cacheManager = FakeImageCacheManager();
+      cacheManager.returns(url, kTransparentImage);
+      final imageAvailable = Completer<void>();
 
-    final ImageProvider imageProvider =
-        CachedNetworkImageProvider(url, cacheManager: cacheManager);
-    final result = imageProvider.resolve(ImageConfiguration.empty);
+      final ImageProvider imageProvider =
+          CachedNetworkImageProvider(url, cacheManager: cacheManager);
+      final result = imageProvider.resolve(ImageConfiguration.empty);
 
-    result.addListener(ImageStreamListener(
-      (ImageInfo image, bool synchronousCall) {
-        imageAvailable.complete();
-      },
-    ));
-    await imageAvailable.future;
+      result.addListener(
+        ImageStreamListener(
+          (ImageInfo image, bool synchronousCall) {
+            imageAvailable.complete();
+          },
+        ),
+      );
+      await imageAvailable.future;
 
-    verify(() => cacheManager.getImageFile(
+      verify(
+        () => cacheManager.getImageFile(
           url,
           key: any(named: 'key'),
           headers: any(named: 'headers'),
           withProgress: any(named: 'withProgress'),
           maxHeight: any(named: 'maxHeight'),
           maxWidth: any(named: 'maxWidth'),
-        )).called(1);
+        ),
+      ).called(1);
 
-    verifyNever(() => cacheManager.getFileStream(
+      verifyNever(
+        () => cacheManager.getFileStream(
           url,
           key: any(named: 'key'),
           headers: any(named: 'headers'),
           withProgress: any(named: 'withProgress'),
-        ));
-  }, skip: isBrowser);
+        ),
+      );
+    },
+    skip: isBrowser,
+  );
 
-  test('Supplying an CacheManager should call getFileStream', () async {
-    var url = 'foo.nl';
+  test(
+    'Supplying an CacheManager should call getFileStream',
+    () async {
+      var url = 'foo.nl';
 
-    var cacheManager = FakeCacheManager();
-    cacheManager.returns(url, kTransparentImage);
-    final imageAvailable = Completer<void>();
+      var cacheManager = FakeCacheManager();
+      cacheManager.returns(url, kTransparentImage);
+      final imageAvailable = Completer<void>();
 
-    final ImageProvider imageProvider =
-        CachedNetworkImageProvider(url, cacheManager: cacheManager);
-    final result = imageProvider.resolve(ImageConfiguration.empty);
+      final ImageProvider imageProvider =
+          CachedNetworkImageProvider(url, cacheManager: cacheManager);
+      final result = imageProvider.resolve(ImageConfiguration.empty);
 
-    result.addListener(ImageStreamListener(
-      (ImageInfo image, bool synchronousCall) {
-        imageAvailable.complete();
-      },
-    ));
-    await imageAvailable.future;
+      result.addListener(
+        ImageStreamListener(
+          (ImageInfo image, bool synchronousCall) {
+            imageAvailable.complete();
+          },
+        ),
+      );
+      await imageAvailable.future;
 
-    verify(() => cacheManager.getFileStream(
+      verify(
+        () => cacheManager.getFileStream(
           url,
           key: any(named: 'key'),
           headers: any(named: 'headers'),
           withProgress: any(named: 'withProgress'),
-        )).called(1);
-  }, skip: isBrowser);
+        ),
+      ).called(1);
+    },
+    skip: isBrowser,
+  );
 
-  test('Supplying an CacheManager with maxHeight throws assertion', () async {
-    var url = 'foo.nl';
-    final caughtError = Completer<dynamic>();
+  test(
+    'Supplying an CacheManager with maxHeight throws assertion',
+    () async {
+      var url = 'foo.nl';
+      final caughtError = Completer<dynamic>();
 
-    var cacheManager = FakeCacheManager();
-    cacheManager.returns(url, kTransparentImage);
-    final imageAvailable = Completer<void>();
+      var cacheManager = FakeCacheManager();
+      cacheManager.returns(url, kTransparentImage);
+      final imageAvailable = Completer<void>();
 
-    final ImageProvider imageProvider = CachedNetworkImageProvider(url,
-        cacheManager: cacheManager, maxHeight: 20);
-    final result = imageProvider.resolve(ImageConfiguration.empty);
+      final ImageProvider imageProvider = CachedNetworkImageProvider(
+        url,
+        cacheManager: cacheManager,
+        maxHeight: 20,
+      );
+      final result = imageProvider.resolve(ImageConfiguration.empty);
 
-    result.addListener(
-        ImageStreamListener((ImageInfo image, bool synchronousCall) {
-      imageAvailable.complete();
-    }, onError: (dynamic error, StackTrace? stackTrace) {
-      caughtError.complete(error);
-    }));
-    final dynamic err = await caughtError.future;
+      result.addListener(
+        ImageStreamListener(
+          (ImageInfo image, bool synchronousCall) {
+            imageAvailable.complete();
+          },
+          onError: (Object error, StackTrace? stackTrace) {
+            caughtError.complete(error);
+          },
+        ),
+      );
+      final dynamic err = await caughtError.future;
 
-    expect(err, isA<AssertionError>());
-  }, skip: isBrowser);
+      expect(err, isA<AssertionError>());
+    },
+    skip: isBrowser,
+  );
 
-  test('Supplying an CacheManager with maxWidth throws assertion', () async {
-    var url = 'foo.nl';
-    final caughtError = Completer<dynamic>();
+  test(
+    'Supplying an CacheManager with maxWidth throws assertion',
+    () async {
+      var url = 'foo.nl';
+      final caughtError = Completer<dynamic>();
 
-    var cacheManager = FakeCacheManager();
-    cacheManager.returns(url, kTransparentImage);
-    final imageAvailable = Completer<void>();
+      var cacheManager = FakeCacheManager();
+      cacheManager.returns(url, kTransparentImage);
+      final imageAvailable = Completer<void>();
 
-    final ImageProvider imageProvider = CachedNetworkImageProvider(url,
-        cacheManager: cacheManager, maxWidth: 20);
-    final result = imageProvider.resolve(ImageConfiguration.empty);
+      final ImageProvider imageProvider = CachedNetworkImageProvider(
+        url,
+        cacheManager: cacheManager,
+        maxWidth: 20,
+      );
+      final result = imageProvider.resolve(ImageConfiguration.empty);
 
-    result.addListener(
-        ImageStreamListener((ImageInfo image, bool synchronousCall) {
-      imageAvailable.complete();
-    }, onError: (dynamic error, StackTrace? stackTrace) {
-      caughtError.complete(error);
-    }));
-    final dynamic err = await caughtError.future;
+      result.addListener(
+        ImageStreamListener(
+          (ImageInfo image, bool synchronousCall) {
+            imageAvailable.complete();
+          },
+          onError: (dynamic error, StackTrace? stackTrace) {
+            caughtError.complete(error);
+          },
+        ),
+      );
+      final dynamic err = await caughtError.future;
 
-    expect(err, isA<AssertionError>());
-  }, skip: isBrowser);
+      expect(err, isA<AssertionError>());
+    },
+    skip: isBrowser,
+  );
 }
