@@ -92,8 +92,10 @@ class TestRenderingFlutterBinding extends BindingBase
 
   @override
   void drawFrame() {
-    assert(phase != EnginePhase.build,
-        'rendering_tester does not support testing the build phase; use flutter_test instead');
+    assert(
+      phase != EnginePhase.build,
+      'rendering_tester does not support testing the build phase; use flutter_test instead',
+    );
     final oldErrorHandler = FlutterError.onError;
     FlutterError.onError = (FlutterErrorDetails details) {
       _errors.add(details);
@@ -109,8 +111,10 @@ class TestRenderingFlutterBinding extends BindingBase
       if (phase == EnginePhase.composite) return;
       pipelineOwner.flushSemantics();
       if (phase == EnginePhase.flushSemantics) return;
-      assert(phase == EnginePhase.flushSemantics ||
-          phase == EnginePhase.sendSemanticsUpdate);
+      assert(
+        phase == EnginePhase.flushSemantics ||
+            phase == EnginePhase.sendSemanticsUpdate,
+      );
     } finally {
       FlutterError.onError = oldErrorHandler;
       if (_errors.isNotEmpty) {
@@ -119,12 +123,14 @@ class TestRenderingFlutterBinding extends BindingBase
           if (_errors.isNotEmpty) {
             _errors.forEach(FlutterError.dumpErrorToConsole);
             fail(
-                'There are more errors than the test inspected using TestRenderingFlutterBinding.takeFlutterErrorDetails.');
+              'There are more errors than the test inspected using TestRenderingFlutterBinding.takeFlutterErrorDetails.',
+            );
           }
         } else {
           _errors.forEach(FlutterError.dumpErrorToConsole);
           fail(
-              'Caught error while rendering frame. See preceding logs for details.');
+            'Caught error while rendering frame. See preceding logs for details.',
+          );
         }
       }
     }
@@ -156,8 +162,9 @@ void layout(
   EnginePhase phase = EnginePhase.layout,
   VoidCallback? onErrors,
 }) {
-  assert(box.parent ==
-      null); // We stick the box in another, so you can't reuse it easily, sorry.
+  assert(
+    box.parent == null,
+  ); // We stick the box in another, so you can't reuse it easily, sorry.
 
   renderer.renderView.child = null;
   if (constraints != null) {
@@ -177,8 +184,10 @@ void layout(
 /// Pumps a single frame.
 ///
 /// If `onErrors` is not null, it is set as [TestRenderingFlutterBinding.onError].
-void pumpFrame(
-    {EnginePhase phase = EnginePhase.layout, VoidCallback? onErrors}) {
+void pumpFrame({
+  EnginePhase phase = EnginePhase.layout,
+  VoidCallback? onErrors,
+}) {
   assert(renderer.renderView.child != null); // call layout() first!
 
   if (onErrors != null) {
@@ -294,8 +303,11 @@ class FakeTicker implements Ticker {
 
   @override
   DiagnosticsNode describeForError(String name) {
-    return DiagnosticsProperty<Ticker>(name, this,
-        style: DiagnosticsTreeStyle.errorProperty);
+    return DiagnosticsProperty<Ticker>(
+      name,
+      this,
+      style: DiagnosticsTreeStyle.errorProperty,
+    );
   }
 }
 
@@ -327,5 +339,6 @@ void expectOverflowedErrors() {
 }
 
 RenderConstrainedBox get box200x200 => RenderConstrainedBox(
-    additionalConstraints:
-        const BoxConstraints.tightFor(height: 200.0, width: 200.0));
+      additionalConstraints:
+          const BoxConstraints.tightFor(height: 200.0, width: 200.0),
+    );
