@@ -1,52 +1,19 @@
+/// Web implementation of CachedNetworkImage
 library cached_network_image_web;
 
 import 'dart:async';
 import 'dart:typed_data';
 import 'dart:ui' as ui;
-import 'dart:ui';
 
 import 'package:cached_network_image_platform_interface'
         '/cached_network_image_platform_interface.dart' as platform
-    show ImageLoader;
-import 'package:cached_network_image_platform_interface'
-        '/cached_network_image_platform_interface.dart'
-    show ImageRenderMethodForWeb;
-import 'package:cached_network_image_platform_interface/types.dart';
+    show ImageLoader, ImageRenderMethodForWeb;
 import 'package:flutter/material.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 
 /// ImageLoader class to load images on the web platform.
 class ImageLoader implements platform.ImageLoader {
-  @Deprecated('use loadBufferAsync instead')
-  @override
-  Stream<ui.Codec> loadAsync(
-    String url,
-    String? cacheKey,
-    StreamController<ImageChunkEvent> chunkEvents,
-    DecoderCallback decode,
-    BaseCacheManager cacheManager,
-    int? maxHeight,
-    int? maxWidth,
-    Map<String, String>? headers,
-    ErrorListener? errorListener,
-    ImageRenderMethodForWeb imageRenderMethodForWeb,
-    Function() evictImage,
-  ) {
-    return _load(
-      url,
-      cacheKey,
-      chunkEvents,
-      decode,
-      cacheManager,
-      maxHeight,
-      maxWidth,
-      headers,
-      errorListener,
-      imageRenderMethodForWeb,
-      evictImage,
-    );
-  }
-
+  @Deprecated('Use loadImageAsync instead')
   @override
   Stream<ui.Codec> loadBufferAsync(
     String url,
@@ -58,22 +25,52 @@ class ImageLoader implements platform.ImageLoader {
     int? maxWidth,
     Map<String, String>? headers,
     ErrorListener? errorListener,
-    ImageRenderMethodForWeb imageRenderMethodForWeb,
-    Function() evictImage,
+    platform.ImageRenderMethodForWeb imageRenderMethodForWeb,
+    VoidCallback evictImage,
   ) {
     return _load(
       url,
       cacheKey,
       chunkEvents,
       (bytes) async {
-        final buffer = await ImmutableBuffer.fromUint8List(bytes);
+        final buffer = await ui.ImmutableBuffer.fromUint8List(bytes);
         return decode(buffer);
       },
       cacheManager,
       maxHeight,
       maxWidth,
       headers,
-      errorListener,
+      imageRenderMethodForWeb,
+      evictImage,
+    );
+  }
+
+  @override
+  Stream<ui.Codec> loadImageAsync(
+    String url,
+    String? cacheKey,
+    StreamController<ImageChunkEvent> chunkEvents,
+    ImageDecoderCallback decode,
+    BaseCacheManager cacheManager,
+    int? maxHeight,
+    int? maxWidth,
+    Map<String, String>? headers,
+    ErrorListener? errorListener,
+    platform.ImageRenderMethodForWeb imageRenderMethodForWeb,
+    VoidCallback evictImage,
+  ) {
+    return _load(
+      url,
+      cacheKey,
+      chunkEvents,
+      (bytes) async {
+        final buffer = await ui.ImmutableBuffer.fromUint8List(bytes);
+        return decode(buffer);
+      },
+      cacheManager,
+      maxHeight,
+      maxWidth,
+      headers,
       imageRenderMethodForWeb,
       evictImage,
     );
@@ -89,11 +86,11 @@ class ImageLoader implements platform.ImageLoader {
     int? maxWidth,
     Map<String, String>? headers,
     ErrorListener? errorListener,
-    ImageRenderMethodForWeb imageRenderMethodForWeb,
-    Function() evictImage,
+    platform.ImageRenderMethodForWeb imageRenderMethodForWeb,
+    VoidCallback evictImage,
   ) {
     switch (imageRenderMethodForWeb) {
-      case ImageRenderMethodForWeb.HttpGet:
+      case platform.ImageRenderMethodForWeb.HttpGet:
         return _loadAsyncHttpGet(
           url,
           cacheKey,
@@ -103,10 +100,9 @@ class ImageLoader implements platform.ImageLoader {
           maxHeight,
           maxWidth,
           headers,
-          errorListener,
           evictImage,
         );
-      case ImageRenderMethodForWeb.HtmlImage:
+      case platform.ImageRenderMethodForWeb.HtmlImage:
         return _loadAsyncHtmlImage(url, chunkEvents).asStream();
     }
   }
@@ -121,25 +117,31 @@ class ImageLoader implements platform.ImageLoader {
     int? maxWidth,
     Map<String, String>? headers,
     ErrorListener? errorListener,
-    Function() evictImage,
+    VoidCallback evictImage,
   ) async* {
     try {
-      await for (var result in cacheManager.getFileStream(url,
-          withProgress: true, headers: headers)) {
+      await for (final result in cacheManager.getFileStream(
+        url,
+        key: cacheKey,
+        withProgress: true,
+        headers: headers,
+      )) {
         if (result is DownloadProgress) {
-          chunkEvents.add(ImageChunkEvent(
-            cumulativeBytesLoaded: result.downloaded,
-            expectedTotalBytes: result.totalSize,
-          ));
+          chunkEvents.add(
+            ImageChunkEvent(
+              cumulativeBytesLoaded: result.downloaded,
+              expectedTotalBytes: result.totalSize,
+            ),
+          );
         }
         if (result is FileInfo) {
-          var file = result.file;
-          var bytes = await file.readAsBytes();
-          var decoded = await decode(bytes);
+          final file = result.file;
+          final bytes = await file.readAsBytes();
+          final decoded = await decode(bytes);
           yield decoded;
         }
       }
-    } catch (e) {
+    } on Object {
       // Depending on where the exception was thrown, the image cache may not
       // have had a chance to track the key in the cache at all.
       // Schedule a microtask to give the cache a chance to add the key.
@@ -149,9 +151,8 @@ class ImageLoader implements platform.ImageLoader {
 
       errorListener?.call(e);
       rethrow;
-    } finally {
-      await chunkEvents.close();
     }
+    await chunkEvents.close();
   }
 
   Future<ui.Codec> _loadAsyncHtmlImage(
@@ -159,7 +160,6 @@ class ImageLoader implements platform.ImageLoader {
     StreamController<ImageChunkEvent> chunkEvents,
   ) {
     final resolved = Uri.base.resolve(url);
-
     // ignore: undefined_function
     return ui.webOnlyInstantiateImageCodecFromUrl(
       resolved,
