@@ -2,40 +2,38 @@
 
 import 'dart:async';
 import 'dart:typed_data';
+import 'dart:ui' as ui;
 
 import 'package:cached_network_image_platform_interface'
     '/cached_network_image_platform_interface.dart' hide ImageLoader;
 import 'package:cached_network_image_web/cached_network_image_web.dart';
+import 'package:file/file.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'dart:ui' as ui;
-import 'package:file/file.dart';
 
 void main() {
-  test('loadAsync returns a stream', () {
-    var imageLoader = ImageLoader();
-    var stream = imageLoader.loadAsync(
-        'test.com/image',
-        null,
-        StreamController<ImageChunkEvent>(),
-        decoder,
-        MockCacheManager(),
-        null,
-        null,
-        null,
-        null,
-        ImageRenderMethodForWeb.HttpGet,
-        () => {});
+  test('loadImageAsync returns a stream', () {
+    final imageLoader = ImageLoader();
+    final stream = imageLoader.loadImageAsync(
+      'test.com/image',
+      null,
+      StreamController<ImageChunkEvent>(),
+      decoder,
+      MockCacheManager(),
+      null,
+      null,
+      null,
+      ImageRenderMethodForWeb.HttpGet,
+      () => {},
+    );
     expect(stream, isNotNull);
   });
 }
 
 Future<ui.Codec> decoder(
-  Uint8List bytes, {
-  int? cacheWidth,
-  int? cacheHeight,
-  bool allowUpscaling = false,
+  ui.ImmutableBuffer buffer, {
+  ui.TargetImageSizeCallback? getTargetSize,
 }) {
   throw UnimplementedError();
 }
@@ -62,14 +60,19 @@ class MockCacheManager implements BaseCacheManager {
   }
 
   @override
-  Stream<FileInfo> getFile(String url,
-      {String? key, Map<String, String>? headers}) {
+  Stream<FileInfo> getFile(
+    String url, {
+    String? key,
+    Map<String, String>? headers,
+  }) {
     throw UnimplementedError();
   }
 
   @override
-  Future<FileInfo?> getFileFromCache(String key,
-      {bool ignoreMemCache = false}) {
+  Future<FileInfo?> getFileFromCache(
+    String key, {
+    bool ignoreMemCache = false,
+  }) {
     throw UnimplementedError();
   }
 
@@ -79,32 +82,45 @@ class MockCacheManager implements BaseCacheManager {
   }
 
   @override
-  Stream<FileResponse> getFileStream(String url,
-      {String? key, Map<String, String>? headers, bool withProgress = false}) {
+  Stream<FileResponse> getFileStream(
+    String url, {
+    String? key,
+    Map<String, String>? headers,
+    bool withProgress = false,
+  }) {
     throw UnimplementedError();
   }
 
   @override
-  Future<File> getSingleFile(String url,
-      {String? key, Map<String, String>? headers}) {
+  Future<File> getSingleFile(
+    String url, {
+    String? key,
+    Map<String, String>? headers,
+  }) {
     throw UnimplementedError();
   }
 
   @override
-  Future<File> putFile(String url, Uint8List fileBytes,
-      {String? key,
-      String? eTag,
-      Duration maxAge = const Duration(days: 30),
-      String fileExtension = 'file'}) {
+  Future<File> putFile(
+    String url,
+    Uint8List fileBytes, {
+    String? key,
+    String? eTag,
+    Duration maxAge = const Duration(days: 30),
+    String fileExtension = 'file',
+  }) {
     throw UnimplementedError();
   }
 
   @override
-  Future<File> putFileStream(String url, Stream<List<int>> source,
-      {String? key,
-      String? eTag,
-      Duration maxAge = const Duration(days: 30),
-      String fileExtension = 'file'}) {
+  Future<File> putFileStream(
+    String url,
+    Stream<List<int>> source, {
+    String? key,
+    String? eTag,
+    Duration maxAge = const Duration(days: 30),
+    String fileExtension = 'file',
+  }) {
     throw UnimplementedError();
   }
 
